SettingsScreen5 ready to go

+++ TouchGFX/gui/src/containers/ScrollSelectedItemContainer.cpp
@@ -10,7 +10,7 @@ void ScrollSelectedItemContainer::initialize()
     ScrollSelectedItemContainerBase::initialize();
 }
 
-void ScrollSelectedItemContainer::updateScrollSelectedItem(int16_t value)
+void ScrollSelectedItemContainer::updateScrollSelectedItem(int value)
 {
     Unicode::snprintf(ScrollSelectedItemBuffer, SCROLLSELECTEDITEM_SIZE, "%d", value);
     ScrollSelectedItem.invalidate();
